refactor(cargo-cost-lint): merge redundant budget_config module into config
module_13.rs (renamed to budget_config.rs) and config.rs turned out to be
the same concern: both parse the budget.toml config table. budget_config.rs
was dead code outside its own tests (its mod was #[allow(dead_code)] and
nothing in main.rs used it), so per issue #400 we merge it into config.rs
rather than keep a second module describing the same config.

The lenient parser is preserved as LenientBudgetConfig with #[allow(dead_code)]
rather than removed, matching the issue's guidance not to delete unused fields
here. No behaviour changes; no module_N identifier remains.
